refactor: resolve comment
