+++ lib/core/data/notification/notification_enums.dart
@@ -1,0 +1,11 @@
+enum NotificationObjectType { 
+  StoreOrder,
+  Event,
+  User
+}
+
+const Map<String, NotificationObjectType> objectTypeMap = {
+  "StoreOrder": NotificationObjectType.StoreOrder,
+  "Event": NotificationObjectType.Event,
+  "User": NotificationObjectType.User,
+};
