+++ lib/core/utils/navigation_utils.dart
@@ -1,23 +1,25 @@
 import 'package:app/core/data/notification/notification_constants.dart';
-import 'package:app/core/domain/notification/entities/notification.dart' as NotificationEntities;
+import 'package:app/core/data/notification/notification_enums.dart';
 import 'package:app/core/utils/dialog_utils.dart';
 import 'package:app/router/app_router.gr.dart';
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
+
 class NavigationUtils {
 
   static handleNotificationNavigate(BuildContext context, String? type, String? objectType, String? objectId) {
-    switch(objectType) {
-      case "StoreOrder": 
+    NotificationObjectType? objectTypeEnum = objectTypeMap[objectType ?? ""];
+    switch(objectTypeEnum) {
+      case NotificationObjectType.StoreOrder: 
         DialogUtils.showAlert(context: context, title: "Redirect to store order");
         break;
-      case "Event":
+      case NotificationObjectType.Event:
         AutoRouter.of(context).navigate(
           EventDetailRoute(eventId: objectId ?? '', eventName: ''),
         );
         break;
-      case "User":
+      case NotificationObjectType.User:
         AutoRouter.of(context).navigate(
           ProfileRoute(userId: objectId ?? '')
         );
